Finish modularization of project
