+++ DungeonProject/DungeonProject/Utilities.py
@@ -1,3 +1,4 @@
+from fbx import *
 import sys
 import random
 
@@ -30,6 +31,62 @@ class Utilities:
         index = next(i for i in range(len(cumulative)) if cumulative[i] > value)
         return data.keys()[index]
 
+    # Creates a box mesh with the specified dimensions and returns a node that contains it
+    @staticmethod
+    def makeBox(width, height, depth, manager, nodeName = "", meshName = ""):
+        width *= 0.5
+        depth *= 0.5
+        newNode = FbxNode.Create(manager, nodeName)
+        newMesh = FbxMesh.Create(manager, meshName)
 
+        #Copy the vertices in the mesh
+        newMesh.InitControlPoints(8)
+        newMesh.SetControlPointAt(FbxVector4(width, 0.0, depth, 0.0), 0)
+        newMesh.SetControlPointAt(FbxVector4(width, 0.0, -depth, 0.0), 1)
+        newMesh.SetControlPointAt(FbxVector4(-width, 0.0, -depth, 0.0), 2)
+        newMesh.SetControlPointAt(FbxVector4(-width, 0.0, depth, 0.0), 3)
+        newMesh.SetControlPointAt(FbxVector4(width, height, depth, 0.0), 4)
+        newMesh.SetControlPointAt(FbxVector4(width, height, -depth, 0.0), 5)
+        newMesh.SetControlPointAt(FbxVector4(-width, height, -depth, 0.0), 6)
+        newMesh.SetControlPointAt(FbxVector4(-width, height, depth, 0.0), 7)
+        newMesh.BeginPolygon(); #bottom
+        newMesh.AddPolygon(0);
+        newMesh.AddPolygon(3);
+        newMesh.AddPolygon(2);
+        newMesh.AddPolygon(1);
+        newMesh.EndPolygon();
+        newMesh.BeginPolygon(); #top
+        newMesh.AddPolygon(4);
+        newMesh.AddPolygon(5);
+        newMesh.AddPolygon(6);
+        newMesh.AddPolygon(7);
+        newMesh.EndPolygon();    
+        newMesh.BeginPolygon(); #front
+        newMesh.AddPolygon(0);
+        newMesh.AddPolygon(4);
+        newMesh.AddPolygon(7);
+        newMesh.AddPolygon(3);
+        newMesh.EndPolygon();    
+        newMesh.BeginPolygon(); #back
+        newMesh.AddPolygon(1);
+        newMesh.AddPolygon(2);
+        newMesh.AddPolygon(6);
+        newMesh.AddPolygon(5);
+        newMesh.EndPolygon();    
+        newMesh.BeginPolygon(); #right
+        newMesh.AddPolygon(0);
+        newMesh.AddPolygon(1);
+        newMesh.AddPolygon(5);
+        newMesh.AddPolygon(4);
+        newMesh.EndPolygon();    
+        newMesh.BeginPolygon(); #left
+        newMesh.AddPolygon(2);
+        newMesh.AddPolygon(3);
+        newMesh.AddPolygon(7);
+        newMesh.AddPolygon(6);
+        newMesh.EndPolygon();
+    
+        #Add the mesh to the node
+        newNode.SetNodeAttribute(newMesh)
 
-
+        return newNode
